Release step: - Docker release and publish - Docker fast-runtime and publish - Generate release with wasm runtime artifact - notify release on slack
Add dry-run option (default no dry-run)

+++ ci/publish-docker-image.sh
@@ -12,6 +12,7 @@ is_a_release="${IS_A_RELEASE:-false}"
 prod_release="${PROD_RELEASE:-false}"
 dev_release="${DEV_RELEASE:-false}"
 test_release="${TEST_RELEASE:-false}"
+fastruntime_release="${FAST_RUNTIME_RELEASE:-false}"
 github_ref_name="${GITHUB_REF_NAME:-}"
 common_file_location="${COMMON_FILE_LOCATION:-not-set}"
 image_artifact=""
@@ -25,6 +26,8 @@ else
   # shellcheck disable=SC1090
   source "${common_file_location}"
 fi
+
+log_info "DRY_RUN: '${DRY_RUN}'"
 
 # Check for command-line options
 while [ $# -gt 0 ]; do
@@ -63,7 +66,12 @@ if [ "${is_a_release}" = "true" ]; then
   docker_tag_full="${github_ref_name}"
 
   log_info "=== Using Docker image artifact from upstream ==="
-  image_name="$(docker load -i "${GITHUB_WORKSPACE}/${image_artifact}.tar" | awk '/Loaded image:/ { print $3 }')"
+  if [ "${DRY_RUN}" != "true" ]; then
+    image_name="$(docker load -i "${GITHUB_WORKSPACE}/${image_artifact}.tar" | awk '/Loaded image:/ { print $3 }')"
+  else
+    log_info "GITHUB_WORKSPACE=${GITHUB_WORKSPACE} image_artifact=${image_artifact}"
+    image_name="__dry_run__"
+  fi
   log_info "=== Loaded already built image '${image_name}' ==="
 
   # Publishing to DockerHub
@@ -79,6 +87,8 @@ if [ "${is_a_release}" = "true" ]; then
     publish_tags=("${docker_tag_full}" "${docker_tag_node}")
   elif [ "${test_release}" = "true" ]; then
     publish_tags=("${docker_tag_full}")
+  elif [ "${fastruntime_release}" = "true" ]; then
+    publish_tags=("fast-runtime")
   fi
 
   # Append -relay to tag names for relay chain images
@@ -91,15 +101,25 @@ if [ "${is_a_release}" = "true" ]; then
 
   for publish_tag in "${publish_tags[@]}"; do
     log_info "Publishing docker image: ${docker_image_build_name}:${publish_tag}"
-    docker tag "${image_name}" "index.docker.io/${docker_hub_org}/${docker_image_build_name}:${publish_tag}"
-    docker push "index.docker.io/${docker_hub_org}/${docker_image_build_name}:${publish_tag}"
+    if [ "${DRY_RUN}" != "true" ]; then
+      docker tag "${image_name}" "index.docker.io/${docker_hub_org}/${docker_image_build_name}:${publish_tag}"
+      docker push "index.docker.io/${docker_hub_org}/${docker_image_build_name}:${publish_tag}"
+    else
+      log_warn "WARNING: 'DRY_RUN' variable is set to 'true'. Don't PUBLISH"
+      log_info "image_name='${image_name}' docker_hub_org='${docker_hub_org}' docker_image_build_name='${docker_image_build_name}' publish_tag='${publish_tag}'"
+    fi
   done
 
   # Extract runtime artifact
   log_info "=== Extract runtime artifact ==="
-  container_id="$(docker create "index.docker.io/${docker_hub_org}/${docker_image_build_name}:${docker_tag_full}")"
-  docker cp "${container_id}":/app/zkv_runtime.compact.compressed.wasm ./zkv_runtime.compact.compressed.wasm
-  docker rm "${container_id}"  # Clean up the container
+  if [ "${DRY_RUN}" != "true" ]; then
+    container_id="$(docker create "index.docker.io/${docker_hub_org}/${docker_image_build_name}:${docker_tag_full}")"
+    docker cp "${container_id}":/app/zkv_runtime.compact.compressed.wasm ./zkv_runtime.compact.compressed.wasm
+    docker rm "${container_id}"  # Clean up the container
+  else
+    log_warn "WARNING: 'DRY_RUN' variable is set to 'true'. CREATE FAKE WASM"
+    echo "index.docker.io/${docker_hub_org}/${docker_image_build_name}:${docker_tag_full} -> WASM" > ./zkv_runtime.compact.compressed.wasm
+  fi
 else
   fn_die "ERROR: the build did NOT satisfy RELEASE build requirements(IS_A_RELEASE variable=${is_a_release}). Docker image(s) was(were) NOT published."
 fi
